slight refactor

+++ app/src/main/java/com/plcoding/soundrecognition/data/models/DataSoundParameters.kt
@@ -2,7 +2,7 @@ package com.plcoding.soundrecognition.data.models
 
 data class DataSoundParameters(val typeName:String, val signalEnvelope:List<Integer>,
                           val rootMeanSquareEnergy:List<Integer>, val zeroCrossingDensity:Integer,
-                          val powerSpectrum:List<Integer>, val spectralCentroids:List<Integer>,
+                          val powerSpectrum:List<Integer>, val spectralCentroids:List<Double>,
                           val spectralFluxes:List<Integer>, val spectralRollOffPoints:List<Double>) {
     override fun toString(): String {
         return "Params(zeroCrossingDensity=$zeroCrossingDensity \n" +
